fix subprocess race condition
