+++ changes/202109021553.bugfix
@@ -1,0 +1,1 @@
+Fixed `Subprocess` race condition reported by golang race detector
